refactor: bring more logic into hookFunc
refactor: bring more logic into hookFunc

fix: rename function name
